Check for environment existence before creating env-pipeline
This PR will add the feature of checking whther the environment
exist or not before creating a environment pipeline

Adds the environment service in serviceFactory

To check it will make a request to env service and get all
the environment in a space. Then conver them into
map[envId]envName and then traverse all the env-pipeline
creation request and check the respective environment
exists or not in map. In case it does not exist, throws error.

Adds test for the change and little bit refactoring.

Fixes https://github.com/fabric8-services/fabric8-build/issues/93
and https://github.com/openshiftio/openshift.io/issues/4566